+++ src/components/project/simulation/geometry/__tests__/index.test.tsx
@@ -84,6 +84,151 @@ describe('components/project/simulation/geometry', () => {
   })
 
   test('with geometry value', async () => {
+    const { unmount } = render(
+      <Geometry
+        geometries={[
+          {
+            id: 'id',
+            name: 'geometry',
+            //@ts-ignore
+            summary: {}
+          },
+          {
+            id: 'id2',
+            name: 'geometry2',
+            //@ts-ignore
+            summary: {}
+          }
+        ]}
+        simulation={{
+          id: 'id',
+          scheme: {
+            category: 'category',
+            name: 'name',
+            algorithm: 'algorithm',
+            code: 'code',
+            version: 'version',
+            description: 'description',
+            //@ts-ignore
+            configuration: {
+              geometry: {
+                index: 1,
+                title: 'Geometry',
+                children: [
+                  {
+                    label: 'Label',
+                    noMeshable: true,
+                    value: 'id'
+                  }
+                ]
+              },
+              boundaryConditions: {
+                index: 1,
+                title: 'Boundary conditions',
+                done: false,
+                dirichlet: {
+                  label: 'dirichlet'
+                }
+              }
+            }
+          }
+        }}
+        setGeometries={setGeometries}
+        swr={swr}
+      />
+    )
+
+    const select = screen.getByRole('combobox')
+    await act(() => fireEvent.mouseDown(select))
+
+    const option = screen.getByText('geometry2')
+    await act(() => fireEvent.click(option))
+
+    await waitFor(() => expect(mockUpdate).toHaveBeenCalledTimes(1))
+    await waitFor(() =>
+      expect(swr.mutateOneSimulation).toHaveBeenCalledTimes(1)
+    )
+
+    unmount()
+  })
+
+  test('with geometry value - wrong geometry', async () => {
+    const { unmount } = render(
+      <Geometry
+        geometries={[
+          {
+            id: 'id',
+            name: 'geometry',
+            //@ts-ignore
+            summary: {}
+          },
+          {
+            id: 'id2',
+            name: 'geometry2',
+            //@ts-ignore
+            summary: {}
+          }
+        ]}
+        simulation={{
+          id: 'id',
+          scheme: {
+            category: 'category',
+            name: 'name',
+            algorithm: 'algorithm',
+            code: 'code',
+            version: 'version',
+            description: 'description',
+            //@ts-ignore
+            configuration: {
+              geometry: {
+                index: 1,
+                title: 'Geometry',
+                children: [
+                  {
+                    label: 'Label',
+                    noMeshable: true,
+                    value: 'id'
+                  }
+                ]
+              },
+              boundaryConditions: {
+                index: 1,
+                title: 'Boundary conditions',
+                done: false,
+                dirichlet: {
+                  label: 'dirichlet',
+                  values: [
+                    //@ts-ignore
+                    {
+                      geometry: 'id'
+                    }
+                  ]
+                }
+              }
+            }
+          }
+        }}
+        setGeometries={setGeometries}
+        swr={swr}
+      />
+    )
+
+    const select = screen.getByRole('combobox')
+    await act(() => fireEvent.mouseDown(select))
+
+    const option = screen.getByText('geometry2')
+    await act(() => fireEvent.click(option))
+
+    await waitFor(() => expect(mockUpdate).toHaveBeenCalledTimes(1))
+    await waitFor(() =>
+      expect(swr.mutateOneSimulation).toHaveBeenCalledTimes(1)
+    )
+
+    unmount()
+  })
+
+  test('with geometry value - error', async () => {
+    // Select error
     mockUpdate.mockImplementation(() => {
       throw new Error('update error')
     })
@@ -124,6 +269,20 @@ describe('components/project/simulation/geometry', () => {
                     value: 'id'
                   }
                 ]
+              },
+              boundaryConditions: {
+                index: 1,
+                title: 'Boundary conditions',
+                done: false,
+                dirichlet: {
+                  label: 'dirichlet',
+                  values: [
+                    //@ts-ignore
+                    {
+                      geometry: 'id2'
+                    }
+                  ]
+                }
               }
             }
           }
@@ -133,7 +292,6 @@ describe('components/project/simulation/geometry', () => {
       />
     )
 
-    // Select error
     const select = screen.getByRole('combobox')
     await act(() => fireEvent.mouseDown(select))